pyproject.toml
[project]
name = "pdf-discovery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "camelot>=12.6.29",
    "camelot-py>=0.11.0",
    "mypy>=1.16.1",
    "pymupdf>=1.26.1",
    "pypdf>=5.6.1",
    "tabula-py>=2.10.0",
    "ultralytics>=8.3.160",
]
